Clean up program startup in Api
